% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Say hi}
\usage{
hello(x)
}
\arguments{
\item{x}{A \code{character(1)} with the name of the person we are saying hi to.}
}
\value{

}
\description{
This function greets someone.
}
\examples{

hello("Mina")
hello("David")
}
